+++ src/main/java/school/faang/user_service/controller/SkillController.java
@@ -1,11 +1,9 @@
 package school.faang.user_service.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.skill.SkillCandidateDto;
 import school.faang.user_service.dto.skill.SkillDto;
-import school.faang.user_service.entity.Skill;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.service.SkillService;
 
@@ -17,7 +15,7 @@ public class SkillController {
     private final SkillService skillService;
 
     public SkillDto create(SkillDto skillDto) {
-        if (validateSkill(skillDto)) {
+        if (skillDto.validateTitle()) {
             return skillService.create(skillDto);
         }
         throw new DataValidationException("skill title is invalid");
@@ -33,9 +31,5 @@ public class SkillController {
 
     public SkillDto acquireSkillFromOffers(long skillId, long userId) {
         return skillService.acquireSkillFromOffers(skillId, userId);
-    }
-
-    private boolean validateSkill(SkillDto skillDto) {
-        return skillDto.validateTitle();
     }
 }
